course_bot/requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
APScheduler>=3.10.0
gspread>=5.10.0
google-auth>=2.0.0
python-dotenv>=1.0.0
pytz>=2023.3
